Move the Auth API methods into a subclass
Change the design to put only generic methods in Net::Duo and to
put each API into a separate subclass.  Each requires separate
keys, so this will make it easier for the user to juggle.  It will
also keep the documentation more focused, since only the methods
from that class will be listed.