Clean up coding style issues
Use a private _fields() method for Net::Duo::Object and ignore the
fact that it appears not to be used.  Ensure all variables in the
SYNOPSIS are declared.  Add more missing stopwords.  Ignore the
if/elsif ladder in Net::Duo::Object for right now pending some
brainstorm about a better way to write it.