Add json methods to every object
Add a json method to every object to return the JSON representation
of the current object data.  We will use this to send data to other
systems.

Currently, this is a bit inefficient since it converts child objects
to JSON and the decodes them and pastes them into the structure so
that the whole object can be converted.  This may be worth another
look later.