Convert validate_passcode into an auth method
Make the major entry point to Net::Duo::Auth a new auth method,
which is like validate_passcode except takes the full range of
arguments as a reference to a hash and supports all of the various
ways of calling auth except for async and requesting sms codes.
Those will be handled in separate methods.